REFACTOR: modify header search input to seperate keyword search hooks
